// src/app/core/enums/color/colorNameEnum.ts
/** 
 * @description Color 中的顏色
 */
export enum ColorNameEnum {
  Gray = 'Gray',
  MidnightBlue = 'MidnightBlue',
  RoyalPurple = 'RoyalPurple',
  SkyBlue = 'SkyBlue',
  Teal = 'Teal',
  ForestGreen = 'ForestGreen',
  Crimson = 'Crimson',
  Coral = 'Coral',
  SunflowerYellow = 'SunflowerYellow',
  Charcoal = 'Charcoal',
  SteelBlue = 'SteelBlue',
  PastelPurple = 'PastelPurple',
  NavyBlue = 'NavyBlue',
  DeepTeal = 'DeepTeal',
  MintGreen = 'MintGreen',
  BurntOrange = 'BurntOrange',
  DarkRose = 'DarkRose',
  Amber = 'Amber',
  LightGray = 'LightGray',
  PowderBlue = 'PowderBlue',
  LightOrchid = 'LightOrchid',
  BabyBlue = 'BabyBlue',
  PaleTurquoise = 'PaleTurquoise',
  MossGreen = 'MossGreen',
  Copper = 'Copper',
  Peach = 'Peach',
  Beige = 'Beige',
  Black = 'Black',
  ElectricBlue = 'ElectricBlue',
  Fuchsia = 'Fuchsia',
  VividCyan = 'VividCyan',
  LimeGreen = 'LimeGreen',
  SpringGreen = 'SpringGreen',
  VividOrange = 'VividOrange',
  CherryRed = 'CherryRed',
  BrightYellow = 'BrightYellow',
}